builder/nat_nv_ingest/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64.0.0", "wheel"]

[project]
name = "nat_nv_ingest"
dynamic = ["version"]
dependencies = [
  "nvidia-nat[all]@git+https://github.com/NVIDIA/NeMo-Agent-Toolkit.git@v1.3.0",
  "nv-ingest-api==2025.10.8.dev20251008",
  "nv-ingest-client==2025.10.8.dev20251008",
  "milvus-lite",
  "minio",
  "pymilvus>=2.5.0",
  "pymilvus.model~=0.3.2",
]
requires-python = ">=3.11,<3.14"
description = "Custom NeMo Agent Toolkit Workflow"
classifiers = ["Programming Language :: Python"]

[project.entry-points.'nat.components']
nat_nv_ingest = "nat_nv_ingest.register"

[tool.setuptools]
packages = ["nat_nv_ingest"]
package-dir = { "" = "src" }

[tool.setuptools.package-data]
nat_nv_ingest = ["configs/*.yml", "configs/*.yaml"]

[tool.uv]
config-settings = { editable_mode = "compat" }
